React changes: Implemented a functionality to delete existing products. it is located at the details view of the product. Add a spinner to indicate users than the delete proccess is running. create modal to confirm if the user really want to perform the delete actions
change the logic of the modifyProduct view a little bit in order to solve problem when trying to make put request by change the method put to post while using axios and then add a " _method " attribute with a value of "put" to the final object that contained the updated information of the product so I could execute the put request correctly.

Add some conditionalls to the client table view so I could render a spinner when loading the clients on the first load of the view, a list of filter results when filtering clients by passing a parameter in the search bar, or just showing a message to indicate that the result was not found.

Custom bootstrap color theme:
Install the bootstrap framework source files so I could use the .scss files and change the default color to some other chose by me. in my case I apply a green scale theme. I have change the variables:  $primary, $dark, $danger, $light, $secondary.

Laravel api changes:
create the update method to change information of an existing product and its image. create a method to delete products by its id.